Add files via upload
𝗧𝗮𝘀𝗸 𝟭: 𝗙𝗿𝗮𝗺𝗶𝗻𝗴 𝘁𝗵𝗲 𝗕𝘂𝘀𝗶𝗻𝗲𝘀𝘀 𝗦𝗰𝗲𝗻𝗮𝗿𝗶𝗼

We are tasked with anticipating the business task that the client might ask. We need to draft questions that we think will be important and relevant to the CEO and CMO. The deliverables are 4 relevant questions each to the CEO and CMO that wants to know the answers to with regards to the dataset provided.

𝗧𝗮𝘀𝗸 𝟮: 𝗖𝗵𝗼𝗼𝘀𝗶𝗻𝗴 𝘁𝗵𝗲 𝗥𝗶𝗴𝗵𝘁 𝗩𝗶𝘀𝘂𝗮𝗹𝘀

With this task, we are tested about what different charts and visuals can be appropriate for different kinds of scenarios. There is a quiz needed to be completed in this task to proceed.

𝗧𝗮𝘀𝗸 𝟯: 𝗖𝗿𝗲𝗮𝘁𝗶𝗻𝗴 𝗘𝗳𝗳𝗲𝗰𝘁𝗶𝘃𝗲 𝗩𝗶𝘀𝘂𝗮𝗹𝘀

In this task, the client provided us with the business tasks, and we have to create visuals based on business scenarios.

KPI's Used-

1- Total Sales
2- Total Revenue
3- No of Products
4-No of Unique Customers
5- Top 10 Products (by Highest Sale & Revenue)
6- Top 10 Products (by Lowest Sale & Revenue)
7- Total Products Sold in Each Month in 2011.
8- Total Revenue Generated in Each Month in 2011.
9-Top 10 Countries by Sales & Revenue
10-Top 10 Customers by Revenue
11-Map Visualization of No of Products Sold in Each Country.





𝗧𝗮𝘀𝗸 𝟰: 𝗖𝗼𝗺𝗺𝘂𝗻𝗶𝗰𝗮𝘁𝗶𝗻𝗴 𝗜𝗻𝘀𝗶𝗴𝗵𝘁𝘀 𝗮𝗻𝗱 𝗔𝗻𝗮𝗹𝘆𝘀𝗶𝘀

We are tasked with creating a presentation that communicates the insights and findings of the analysis. Everything done from start to finish is summarized and added to present to the client